+++ src/themes/blueberry/Table.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export default props => <table className="bb-table" {...props} />
