+++ src/themes/blueberry/Wrap.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export default ({ style = {}, ...props }) => <b style={{ backgroundColor: 'yellow', ...style }} {...props} />
